% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_decompress.R
\name{wpd_decompress}
\alias{wpd_decompress}
\title{wpd_decompress}
\usage{
wpd_decompress(page_views, year, page_id, lang, page_name)
}
\arguments{
\item{page_views}{single cell from column of same name from db query result}

\item{year}{single cell from column of same name from db query result}

\item{page_id}{single cell from column of same name from db query result}

\item{lang}{single cell from column of same name from db query result}

\item{page_name}{single cell from column of same name from db query result}
}
\description{
wpd_decompress
}
