% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_dump_path.R
\name{wpd_dump_path}
\alias{wpd_dump_path}
\title{wpd_dump_path}
\usage{
wpd_dump_path(path)
}
\arguments{
\item{path}{path to set - location of Wikipedia dump SQLite databases}
}
\description{
Store and retrieve path to Wikipedia dump SQLite databases.
}
\examples{

# setting path - #1
options("./")

# setting path - #2
wpd_dump_path("./")

# getting path - #1
options("wpd.path")

# getting path - #2
wpd_dump_path()

}
