% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_languages.R
\docType{data}
\name{wpd_languages}
\alias{wpd_languages}
\title{wpd_languages}
\format{An object of class \code{character} of length 20.}
\usage{
wpd_languages
}
\description{
wpd_languages
}
\keyword{datasets}
